+++ packages/board-server/src/server/common.ts
@@ -6,8 +6,9 @@ import type { ViteDevServer } from "vite";
 import { existsSync } from 'fs';
 
 import type { ServerConfig } from "./config.js";
-import { notFound } from "./errors.js";
+import { notFound, serverError } from "./errors.js";
 import type { PageMetadata } from "./types.js";
+import { createReadStream } from "fs";
 
 const MODULE_PATH = dirname(fileURLToPath(import.meta.url));
 
@@ -69,12 +70,23 @@ export const serveFile = async (
   const contentType = CONTENT_TYPE.get(extname(path)) || DEFAULT_CONTENT_TYPE;
   try {
     const resolvedPath = resolve(root(), path.replace(/^\//, ''));
-    console.log(`Resolved path: ${resolvedPath}`);
-    console.log(`File exists: ${existsSync(resolvedPath)}`);
-    let contents = await readFile(resolvedPath, "utf-8");
-    if (transformer) contents = await transformer(contents);
-    res.writeHead(200, { "Content-Type": contentType });
-    res.end(contents);
+    if (contentType.startsWith("text/") || contentType === "application/json") {
+      let contents = await readFile(resolvedPath, "utf-8");
+      if (transformer) contents = await transformer(contents);
+      res.writeHead(200, { "Content-Type": contentType });
+      res.end(contents);
+    } else {
+      return new Promise<void>((resolve) => {
+        const fileStream = createReadStream(resolvedPath);
+        res.setHeader("Content-Type", contentType);
+        fileStream.pipe(res);
+        fileStream.on("error", () => {
+          notFound(res, `Static file not found`);
+          resolve();
+        });
+        fileStream.on("end", resolve);
+      });
+    }
   } catch (err) {
     console.error(`Error serving file: ${err}`);
     notFound(res, "Static file not found");