+++ packages/board-server/src/server/cors.ts
@@ -4,24 +4,23 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import type { IncomingMessage, ServerResponse } from "http";
-import { getStore } from "./store.js";
+import type {
+  IncomingMessage,
+  OutgoingHttpHeaders,
+  ServerResponse,
+} from "http";
 
-const CORS_HEADERS = {
+const CORS_HEADERS: OutgoingHttpHeaders = {
   "Access-Control-Allow-Origin": "*",
   "Access-Control-Allow-Methods": "OPTIONS, POST, GET, DELETE",
   "Access-Control-Allow-Credentials": "true",
   "Access-Control-Allow-Headers":
     "Content-Type, Access-Control-Allow-Headers, Authorization",
   "Access-Control-Max-Age": 2592000, // 30 days
-} as Record<string, string | number>;
-
-// Read once from the database and cache it.
-// TODO: Make this a bit more dynamic.
-const CONFIG = await getStore().getBoardServerCorsConfig();
+};
 
 export const corsAll = (req: IncomingMessage, res: ServerResponse) => {
-  const headers = structuredClone(CORS_HEADERS);
+  const headers = { ...CORS_HEADERS };
   headers["Access-Control-Allow-Origin"] = req.headers.origin || "*";
 
   if (req.method === "OPTIONS") {
@@ -32,9 +31,11 @@ export const corsAll = (req: IncomingMessage, res: ServerResponse) => {
 
   const method = req.method || "GET";
   if (["GET", "POST"].indexOf(method) > -1) {
-    Object.entries(headers).forEach(([key, value]) => {
-      res.setHeader(key, value);
-    });
+    for (const [key, value] of Object.entries(headers)) {
+      if (value !== undefined) {
+        res.setHeader(key, value);
+      }
+    }
     return true;
   }
 
@@ -43,17 +44,18 @@ export const corsAll = (req: IncomingMessage, res: ServerResponse) => {
   return false;
 };
 
-export const cors = (req: IncomingMessage, res: ServerResponse) => {
+export const cors = (
+  req: IncomingMessage,
+  res: ServerResponse,
+  allowedOrigins: Set<string>
+) => {
+  const headers = { ...CORS_HEADERS };
   const origin = req.headers.origin || "";
-  const isLocalhost = origin.includes("localhost");
   const host = req.headers.host || "";
   const sameOrigin = origin.includes(host);
-  const headers = structuredClone(CORS_HEADERS);
+
   const isAllowed =
-    isLocalhost ||
-    sameOrigin ||
-    origin.length === 0 ||
-    CONFIG?.allow?.includes(origin);
+    sameOrigin || origin.length === 0 || allowedOrigins.has(origin);
   if (!isAllowed) {
     res.writeHead(403);
     res.end(`${origin} is not allowed for the request.`);
@@ -70,9 +72,11 @@ export const cors = (req: IncomingMessage, res: ServerResponse) => {
 
   const method = req.method || "GET";
   if (["GET", "POST"].indexOf(method) > -1) {
-    Object.entries(headers).forEach(([key, value]) => {
-      res.setHeader(key, value);
-    });
+    for (const [key, value] of Object.entries(headers)) {
+      if (value !== undefined) {
+        res.setHeader(key, value);
+      }
+    }
     return true;
   }